Provide cohort on creating a new declaration
As `cohort` is a required attribute we need to provide it when we create a
declaration in the `RecordDeclaration` service.

We don't, however, want to match on `cohort` when finding existing declarations
(to keep the behaviour consistent with how it is currently).

Split the find/create logic to find without cohort and create with cohort.