Update cohort change logic to use payments_frozen_at field
When these methods were initially written we assumed that if a cohort was 3
years older than the current/latest cohort it could be considered 'closed'.

The schools team have since added a dedicated flag to `Cohort` to indicate that
it is closed/has had payments frozen.

Update the `eligible_to_change_cohort_and_continue_training` and `archivable`
methods to use this flag instead of the 3 years previous logic.