Align cohort change logic with schools team
Tweaks the logic to change cohort and continue training to aign with the
schools team. This means instead of checking for +/- 3 years from their current
cohort, we will instead let them transfer only ot the active registration
cohort and only back to a cohort that is payments frozen and for which they
have billable declarations.