Expose previous, frozen cohort to lead providers
When a participant is migrated from a payments frozen cohort, we want to expose
this to lead providers in the participant response.

Add `previous_payments_frozen_cohort` field to participant serializers that
will be `nil` if the participant has not migrated cohort due to payments being
frozen and will be the cohort start year of the frozen cohort if they have been
migrated.

It will perform an extra query for these participants, but given the number is
going to be low it shouldn't have a major impact on performance.