Update ParticipantPresenter to use previous_payments_frozen_cohort
This was added for the serializers and we may as well use it here as well. It
also contains an extra check to make sure the previous cohort is frozen, in
case the participant moved cohort using the normal schedule change and then
migrated cohort due to payments being frozen.