Fix broken specs in main
The `change_cohort_and_continue_training_support` specs were failing due to the
milestone date and start date being set to the same value when we're at the end
of the month. Updating the spec to use the `start_date` rather than adding
`1.day` fixes it (as its greater than or equal to).

The cohort changed error message appears to have changed in `main` since the
change schedule for muti cohorts PR was raised and it was failing after being
merged; this updates the error message to match.