[CPLD-3077] Add funded place to previously funded logic
If funded place is set to true, it should be factored into the the
previously funded logic. If it's not set we default to only
relying on `eligible_for_funding`.

Add tests and ammend logic everywhere we calculat previously funded.
This is including the ineligible reason as well as the funding eligibility
service used by the NPQ app