[CPLD-3077] Add flag to include funded_place in eligibility calculation
The eligibility method is used in the API to surface if an application is
eligible for funding. We cannot change that logic to include funded place.
So it should remain the same, however previously funded will include it
because we have to calculate previously funded applications in the new way.

However the fundable method used in recording declarations needs to take into
account the new logic, so add a flag only there to account for funded place.