Ensure next output fee statement exists when attaching declaration
Currently, if we attempt to attach a declaration when there are no output fee
statements available it will result in a 500 error. Instead, we want to surface
a helpful error message to lead providers encountering this scenario.

Update `RecordDeclaration` and `ClawbackDeclaration` to ensure the next output
fee statement exists. Originally I tried to add this validation to the
`StatementAttacher` and bubble it up from there, but we can't do that without a
substantial refactoring due it being called _after_ the `RecordDeclaration`
validation as part of the transaction (if we call it during the validation it
would result in the declaration being created prematurely). I also attempted to
extract this into a custom validator, but the differences between the two
services that call the statement attacher means this isn't viable/clean.

The `StatementAttacher` is also called as part of
`RecordDeclarations::Actions::MakeDeclarationsEligibleForParticipantProfile`,
however this doesn't appear to be exposed to lead providers/handle errors in
general, so I've left as-is.