Use declaration cohort when attaching to statement
Currently, we look to the latest induction record for the lead provider to find
the cohort when attaching a declaration to a statement. This is no longer
correct for participants that have declarations in 2021 and have since been
moved to 2024. It would result in the declaration being voided against a 2024
statement instead of the original 2021 statement.

To correct this behaviour we can take the `declaration.cohort` when attaching
to a statement, which will always be the cohort that the declaration was
initially created in.