Relax restrictions on MigrateDeclarationsBetweenStatements
Up until now we have restricted the usage to only statements that are future
dated. This is too restrictive for a new use case, so we are relaxing it to
migrate statements that are not future dated (albeit with a warning).

We also want to remove the requirement of having to specify an output_fee for
the from statement after the migration has completed. If we find we want to
modify the from statement again in the future we could re-introduce in a more
flexible way, such as with a `from_statement_updates` hash (similar to the
`to_statement_updates` already supported).