Amend sorting in NPQ participants query
Currently the default sorting looks into the nested npq_profiles.
Without pagination this works ok, however with pagination, pagy
includes the timestamp on the profiles in the limit/offset distinct
query, which means that if a user has multiple profiles with
different timestamps we return duplicate ids instead of the real ids
for users.

Instead sort by user created at to avoid any timestamps being included
in the pagy queries when calculating which ids to offset and return.

The test fails with the previous sorting but doesn't with the current
one.