Ignore declarations in frozen cohort on ChangeSchedule
For participants that have been migrated out of a payments frozen cohort, we
want to ignore any declarations made in a frozen cohort when determining if
they can change schedule.

This will allow providers to migrate participants to an extended schedule
after they have migrated from the 2021 payments frozen cohort to the 2024
active registration cohort, even if they have billable declarations in the
2021 cohort.

They will not be able to change the cohort, only the schedule.