Allow providers to migrate participants from 2021 to 2024
The `ChangeSchedule` service already has this functionality, however it was
locked down so that providers could not perform the action. It has been
decided we want to open up this functionality to LPs going forward, so we
can do this by passing `allow_change_to_from_frozen_cohort: true` into the
`ChangeSchedule` service.