Add initializer to remove NPQ references from swagger docs
We want to remove any NPQ references from the swagger docs in environments
where we are running the dedicated NPQ lead provider API.

As the swagger JSON files are no longer generated by `rswag` we need to cut
the NPQ references out of the files directly.

Add initializer that will strip all references of NPQ from the swagger docs.
It performs a deep-search of the JSON structure for references to `npq` and
purges them.