Switch to Middleman extension
The Swagger docs are generated using Middleman; modifying the JSON files on
Rails boot is too late. Instead, we need to modify them when the `middleman
build` command runs. There is an existing extension we tap into in otder to do
this.

Update the deploy CI so build a second docker image that has NPQ swagger docs
removed, so that we can deploy it to a specific environment.