Make change lead provider controller handle both suport queries
This commit moves the controller to the parent module as it will handle wizard flow for support queries for changing lead provider for both cohorts and participants.
Removes the intro step as this has different content and is handled by a normal action and conditional rendering.
Presence of params[:participant_id] determines which flow the user will use.