Add separation environment
We want to add a new `separation` environment alongside the existing `sandbox` environment that can be used by providers to test a version of ECF with all the NPQ functionality removed/disabled.

This commit covers:

- Adding the `separation` environment (based off `sandbox`)
- Adding `Makefile` support
- Deploy on merging to `main`
- Seed data
- Enable time traveller for testing
- Add Rails credentials for new env

I have also setup a new BigQuery events dataset for this environment and it has a new key vault copied from `sandbox`.