Remove NPQ references from JSON swagger files
In addition to scrubbing NPQ references from the guidance/API docs, we also
want to remove them from the underlying JSON files.

Add `Preprocessor` to iterate over and remove NPQ references from JSON files
when `REMOVE_NPQ_REFERENCES` is `true`.

For simplicity and given that this isn't going to stay in the codebase forever,
I've opted to do a VCR-style test to check the before/after from running the
spec files through the preprocessors.