Add support and links to api docs footer
Add the support paragraph/email and links to the privacy policy, cookies
and accessibility to the footer.

The links are supported by the tech docs gem, however the support paragraph
required overriding the footer from the template. Usually in a Middleman
template you can just specify a partial in the correct place and it will
use it/override for you, however for some reason this doesn't work with the
tech docs template. I think because of the way it builds/pulls from the
tech docs gem. Instead, I've had to copy the layouts to our repository,
rename them and update all pages to use them by default. We can then
specify a custom footer.