Add validation to ensure NPQContract/Statement is present
In the accept, change schedule and change funded place actions we want to
ensure the NPQContract and (for some) the Statement is present when the
`npq_capping` feature is enabled.

Add validation that raises a
`Api::Errors::MissingNPQContractOrStatementError`, to be rescued in the API
controller and an error message returned.