Focus on continue button after selecting file
On iOS using VoiceOver when selecting a file on the school report CSV upload
page the focus resets to the first item on the page (skip to main content).

I can't find anything specific to our web app that would cause this, so I think
its default iOS behaviour after a file picker modal appears and then Safari
re-gains focus.

To make it more intuitive for the user I've added an `onchange` that will focus
on the form submit button once a file has been selected (I tried to refocus on
the file input but this doesn't work for some reason).

A side-effect is that this will change for all users; so on desktop the file
input would retain focus by default but now it will move to the continue
button. Short of adding a voice over/mobile detection library to make this more
specific there's not much we can do here.