Add rake task to bulk mark applications as pending
Add a new one off service to make a group of applications `pending`. Add rake
task so that we can call the service with a CSV of application ids and log out
the updates (changed to pending, already pending or an error preventing it from
moving to a pending state).