Don't make `awaiting_clawback` line items `payable` on migrating declarations
We have a bug when migrating declarations to a `Payable` statement whereby we will put the declaration through the `MarkAsPayable` service regardless of its current state. This results in `awaiting_clawback` declarations being made `payable`.

Instead, we want to follow the logic in the rest of the service and only make `eligible` declarations `payable` when moving to a `Payable` statement.

As it stands, the following behaviour will apply for declarations moving to a `Payable` statement:

- eligible -> payable
- awaiting_clawback -> awaiting clawback
- payable -> payable
- submitted/voided/ineligible (ignored - no line items)
- paid/clawed_back (raise error, can't migrate from paid statement)