Fix flakey test
The test for the `awaiting_clawback` declaration being migrated is failing in CI due to the statement for the `paid` and `awaiting_clawback` line items having the same name. To ensure only the payable statement (with the `awaiting_clawback` declaration) is used in the migration we give it another name.