Generate known tokens for parity check on migration DB restore
We hit an issue when running the original token provider whereby the keys would
generate but not work in the migration environment. This was due to the keys
being generated using the NPQ reg migration environment secret_key_base but
verified in ECF using a different secret_key_base.

To get around this, we are going to use known tokens; each migration keyvault
has the same set of tokens indexed by the lead provider `ecf_id`. After
restoring the migration database we will create the known-tokens for the parity
check service to use.