Clean up maintenance banner
Clean up the maintenance banner to match the latest design.

Fix broken `Dismiss` button; swap to a link to be consistent with the GOV.UK
banner component usage.

Change route to GET so we can use it from a regular link in the GOV.UK
components helpers (not ideal but allows us to use the component to render the
banner).

Add missing test coverage.