Switch API doc NPQ removal to feature flag
Currently we remove NPQ references on the separation environment; going forward
we want to remove all references when the disable_npq feature is enabled.

We will need to reboot the app instance to get the correct API docs as we
determine this in an initializer on boot.

Only run initializer when ran in the context of a `rails server` so that we
don't get database connection errors when pre-compiling assets. Move to
after_initialize to ensure the autoloading has finished and FeatureFlag is
defined/loaded. We also need to wait for/ensure the DB is connected so this
initializer can check the feature state.