Relax `UnfundedMentorsQuery`
To be consistent with the current query and return the same users we need to
include `InductionRecord` instances for the mentor where they have a challenged
partnership or no lead provider. We also need to `LEFT OUTER JOIN` on the
`participant_identities` as some users do not have a `preferred_identity` on
the `InductionRecord` that still exists.

There are also some profiles that don't have any `induction_records` so we need
to `LEFT OUTER JOIN` on those as well.