Add callback to ensure `ParticipantDeclaration` type consistency
A `ParticipantProfile` shouldn't change from an `ECT` to a
`Mentor`, but in case someone does this manually via the rails
console we want to ensure consistency on the declaration types.

Change the `participant_declarations` to the correct type when
the `ParticipantProfile#type` is changed.

Ensure the declaration is not touched when this happens.