Revert end to end tests to not use Knapsack
The end-to-end tests are setup differently; instead of running a subset of
tests on each worker, each tests is ran on all workers and we instead split the
scenario test data across the workers, so each one runs all tests for a subset
of the scenario data.

There's not an easy way to optimise this using the Knapsack gem, so we are
reverting to the previous method.