Add and maintain `temp_type` on `ParticipantDeclaration::ECF`
We are splitting `ECF` declarations into `ECT` and `Mentor` declarations. As
calculating the declaration type takes a few minutes, in order to do it without
downtime or data errors we are going to add a `temp_type` column that we can
populate and maintain ahead of the switch.

Once shipped, we can then switch the data of `new_type` to `type` quickly and
start referencing the new column.