Temp manually move transfers in swagger file
In the current swagger file the transfers endpoints are between
`participants/ecf` and `/participants/ecf/{id}/defer`. This is to get the
correct ordering in the UI, but isn't achievable as it involves mixing the
rswag specs - we would have to combine `ecf_participant_transfers_spec` into
`ecf_participants_spec`.

This isn't a problem when we switch to manually sorting the UI in the front
end, this commit is just to bring them back inline for comparisons sake.