Fix 500 error on change schedule endpoint
We have noticed lead providers getting a 500 error on the `change-schedule`
endpoint when they attempt to change the schedule of a participant not
associated with them via their induction records (so
`relevant_induction_record` is `nil`).

On a normal change schedule operation one or more validators will catch this
and return an error to the lead provider. If they are changing schedule due to
payments being frozen, however, and the participant has no billable
declarations, then the error occurs when the code attemps to determine the
`current_cohort` of the participant (as it does this via
`relevant_induction_record`).

To bring the behaviour for changing cohort due to payments frozen inline with
the normal change schedule operation we can instead fail silently on
`current_cohort`, returning `nil` if there is no `relevant_induction_record`
and thereby let one of the follow up validations catch and raise a bad request
error.

Add test case to cover this scenario explicitly.