Switch from Selenium to Cuprite
We have noticed that our CI is hanging randomly on the feature specs that are
using the Javascript Selenium webdriver. It appears to be a HTTP timeout that
Selenium experiences, though its not easy to replicate or debug.

In an effort to fix the issue we can move to Cuprite, which is a pure Ruby
driver for Capybara. An issue with doing this is that the latest versions of
`axe-core-api` have not been updated for Cuprite support (they expect
Selenium). It does, however, work with versions prior to `4.3` so we can pin
the version.

This should get rid of the flakiness as its removing Selenium, which is where
it falls down. I'm still not clear on why Selenium is flakey though.