Split feature specs in CI and fail fast
So that we aren't waiting a long time to see a failure when we get a random
Net:HTTP timeout issue, we can split the feature specs and pass `--fail-fast`
so that the job will fail on the first feature spec that fails. Otherwise we
can end up waiting a long time for each test to fail and as a result the
workflow won't fail for over an hour.