Send correct school name in withdrawn participant email
When we withdraw a participant we send an email with details of the school they
were withdrawn from. Currently, however, we take the participant profile ->
school, which is not always the school relevant to the withdraw action. We
should instead be taking the school from the relevant induction record, in case
the participant has transferred schools in the past.

Update `WithdrawParticipant` action to take the school from the
`relevant_induction_record` instead of the `participant_profile`.