Allow `anyOf` on `data` attributes in API docs
Currently, if you set a `data` attribute up with an `anyOf` value set it will
output the "It conforms to" text as well as the "This conforms to" text. This
is due to `data` being a special attribute whereby we usually link to another,
single schema.

To get around this and allow `anyOf` for `data` attributes we only want to
output the "It conforms to" links if there is _not_ an `anyOf` value set
present.