Update POST /api/participant-declarations API docs for mentor funding
The API docs for creating a declaration need to change to inform lead providers
that they will only be able to use certain payloads depending on the cohort
that is being referenced and the type of participant in question.

For pre-2025 cohorts the request schemas remain the same (started, retained,
extended and completed). For 2025 and later when a declaration is being made
for a mentor we will only accept started and completed.