Expose uplift payable in assurance report CSV
We currently expose the `pupil_premium_uplift` and `sparsity_uplift` attributes
in the assurance report CSVs, however we want to add an additional `Uplift
Payable` column in the CSV, which is `true` if the participant qualifies for
uplift payments (if either `pupil_premium_uplift` or `sparsity_uplift` are
`true`).

Add test coverage for `CsvSerializer`.