Remove unused cookie session store from `production.rb`
We define a `cookie_store` in `production.rb`, however this is later overriden
by the `session_store.rb` initializer, which specifies an
`active_record_store`.

It has been confirmed on the production rails console that the active record
store is in use:

```
irb(main):001> Rails.application.config.session_store
=> ActionDispatch::Session::ActiveRecordStore
```

The cookies have also been checked in production to ensure the
`_early_career_framework_session` cookie key is not present (only the
`_ecf_session` which relates to the `active_record_store`).