Add report-only Content Security Policy
We want to re-introduce the Content Security Policy, however we want it to be
configured to report only initially so that we can monitor for any violations
before enforcing it.

Add CSP headers and re-implement security headers without the `secure_headers`
gem.

Implement separate, locked down CSP for API endpoints.

Update inline scripts and styles to use a nonce.