Add Content Security Policy header to API documentation
As the API documentation is statically compiled into the `/public` directory of
the Rails app it bypasses the Rails middleware and CSP header we serve in the
rest of the application.

Add `meta` tag to specify a Content Security Policy for the documentation. We
allow `unsafe-inline` styles for the `rogue` gem syntax highlighting CSS.

Shifts the inline Javascript into `application.js` so we don't have to allow
`unsafe-inline` scripts.