Fix CSP report-uri causing feature spec failures
When we return a `head :ok` on the CSP report-uri response it causes a couple
of feature specs to flake out. I have absolutely no idea why this is.

The trigger seems to be an `eval` call in one of our dependencies that is
marked as blocked even when we have `unsafe-eval` in the `script_src`
directive. Apparently modern browsers can choose to ignore `unsafe-eval` even
when its in the directive.

A fix for this is to reutrn a 204 no content, which is an acceptable response
to the `report-uri` call. Again, I have no idea why this lets the test pass.