Add clawed back count to finance statements
We want to display the clawed back declarations couynt in the finance statement
summary.

The clawed back declarations are fudged due to the factoy not playing ball;
getting them on the November statement was turning into a lot of test setup, so
I opted to manually transition submitted declarations (the count functionality
is tested indepndently in the calculator).

We also want to always show the extended declaration counts, even when 0.