Exclude 2025 and later participants from delivery partners
We want to exclude participants in the 2025 cohort from the delivery partners
participants view. We also want to hide the 2025 and later cohorts from the
academic year dropdown.

This is indended to encourage lead providers to carry out their contractual
duties and ensure there is no ownness on the DfE to share details of
participants in 2025 and later cohorts.