Use correct serializer on v2 POST declarations endpoint
We pass an explicit v1 serializer on the create action for declarations which
is inherited in the v2 controller. For other actions we override with the
`serializer_class` hook. This results in the wrong data being returned for the
v2 create declarations endpoint.

Ensure correct serializer is used in v2 create declarations action.

Add test coverage to prevent this kind of regression in the future.