Add `voided_at` and `voided_by_user` to `ParticipantDeclaration`
We want to maintain a record of who voided a declaration when it happens
through the user interface.

Add `voided_at` and `voided_by_user` to `ParticipantDeclaration` - these will
be `nil` when a lead provider has voided the declaration via the API.

The `voided_at` column appears to have been part of the model years ago, which
is why it was later ignored - see #2320 - reversing that change here as we
re-introduce it.