Update `VoidParticipantDeclaration` to support `voided_by_user`
When we void a declaration manually via the admin interface we want to track
the user that performed the void and the time that it was voided. We have
`voided_at` and `voided_by_user` as attributes on the `ParticipantDeclaration`
model that we can use.

Add a `voided_by_user` required parameter to the `VoidParticipantDeclaration`
service initializer and update current usage of the service to pass `nil` (to
retain current behaviour).

When we come to expose this in the admin interface we can pass the currently
signed in user. If a user is provided we will populated `voided_by_user` and
set `voided_at` to the curent time.