Add ability to void declarations in finance dashboard
We want to be able to void declarations in the finance dashboard.

Add `Void` link to the participant declarations if they are in a voidable state
(if paid they must also have an output fee statement available).

Add summary page to detail declaration prior to voiding.

On voiding, redirect to the participant page with a success message.

If there is an error when voiding, redirect to the participant page with a
failure message.