Add service to mark CallOffContract records as unused
Due to historical/manual changes to contract data over the years we are in a
position where we have a number of contracts that are either duplicates or are
unused (i.e. not referenced by any financial statements).

Add `Oneoffs::MarkContractsAsUnused` service to find unused/duplicated
contracts and prefix their version with `unused_`. This will enable us to
identify them and exclude them from the data migration to ECF 2.0.

We identify orphaned contracts by finding those without a corresponding
statement (matching on lead provider, cohort and contract version).

Duplicate contracts are initially identified by comparing lead provider, cohort
and version and we then perform a deep-comparison of the `CallOffContract` and
`ParticipantBand` attributes.

Any contracts that match on lead provider/cohort/contract version but differ
when we perform the deep comparison are not marked as unused and we instead
record how many of these we encounter during the process.