Add `not_flagged_as_unused` scope to `CallOffContract`
We will be referencing the used/unused status of call off contracts outside the
oneoff service to prefix them as `unused_`, so it makes sense to keep the scope
and terminology in the model itself.

I expect when we come to migrate the contract data we will add a
`not_flagged_as_unused` scope to get the contracts that are applicable to
migrate over.