Optimize ParticipantsFilter query
Since updating the `ParticipantsFilter` to filter on cohort by default (to
exclude 2025 and later) we have noticed timeouts on the corresponding page in
the dashboard.

It looks like when we join on `cohort` it ends up pulling in a bunch of other
tables (as it has to go through induction programme -> school cohort -> cohort)
and the query that active record ends up with is quite large and slow.

To avoid this, we can instead use a subquery to find the induction programmes
that are associated with relevant cohorts and then filter on those, avoiding
the many joins.

This appears to be more performant; the underlying query is much smaller and on
doing some contrived performance testing against sandbox it seems to take less
time:

- Without cohort filtering; ~2.3s
- With joining on cohort to filter: ~3.2s
- Using a subquery on induction programme: ~2.5s