Prevent mentors changing cohort due to payments frozen
Currently we allow both ECTs and mentors to change schedule from a payments
frozen cohort to the active registration cohort if they have billable
declarations in the frozen cohort.

Going forward we will be marking these mentors as complete which will exclude
them from being eligible to transfer from a payments frozen cohort, but for a
belt and braces approach we also want to guard against this explicitly in the
`ChangeSchedule` service.

Add guard clause to prevent mentors from transferring due to payments being
frozen. We will still allow transferring out of a payments frozen cohort if
there are no billable declarations against the participant (though they won't
be flagged as `cohort_changed_after_payments_frozen` in this scenario).