Fix n+1 on CSV download of delivery partner participants
The `DeliveryPartners::InductionRecordsSerializer` loads the TRN of the
participant via the induction record -> participant profile -> teacher profile,
resulting in an n+1 as we don't preload this association.

Preload the participant profile -> teacher profile.

The user -> teacher profile is still needed for the regular web view to avoid
an n+1 there.