Change wording of error when already clawing/clawed back
The error message we serve to lead providers when they try and clawback a
declaration that is either `clawed_back` or `awaiting_clawback` already is not
clear.

Update the error to be clearer and add test coverage.

I checked to see if any states that aren't `clawed_back` or `awaiting_clawback`
could end up hitting this error message, but the only other state that gets
here is `voided` and that is picked up by the earlier check and raise.