Only use Induction::SetCohortInductionProgramme when needed
Attempting to run the closing cohort seed on sandbox was taking a very long
time; it boils down to using this service to setup the induction programme. It
ends up calling `MigrateParticipantsToNewProgramme` to migrate all the IRs when
a default induction programme is already present.

Instead, we only want to call it when there isn't a default induction programme
on the school cohort.