Clear evidence_held if not applicable when recording declaration
The `EvidenceHeldValidator` will only validate the evidence held if it applies
to the declaration being recorded.

It applies for detailed evidence types if the declaration type is not `started`
or the `evidence_held` value is provided.

It applies for normal evidence types if the declaration type is not `started`.

This means if the declaration is started and for normal evidence types and the
`evidence_held` value is provided it will not be validated but it will end up
being persisted against the declaration.

We can avoid this by either returning a new error message when an
`evidence_held` is provided or by sanitizing the field as part of the
validation process (clearing it if it is not required).

I opted for the latter in this commit in order to keep the behaviour of the API
consistent with what lead providers expect at the moment. Its not ideal as the
validator shouldn't be modifying the value in my opinion, but its the most
pragmatic solution (a better solution would require a larger refactoring of the
validator).