Run analytics job `after_commit` instead of `after_save`
There is a race condition when the various models that have an
`update_analytics` callback are persisted in a transaction (as is the case with
some of the seed data).

The `after_save` will enqueue the job before the transaction is commited; if a
worker picks up the job it attempts to deserialize and verify the model
argument and raises an error as it hasn't been commited to the database yet.

Using `after_commit` instead of `after_save` ensures the model will be in the
database. As all changes aren't readily available to `after_commit`
(`previous_changes` only retains changes from the last save, which is often
empty) we are leveraging `ar_transaction_changes`, which retains all changes
from a commit across multiple saves.