Avoid serializing entire objects for job arguments
We currently pass in (and therefore serialize) whole objects for the various
analytic job parameters. This is inefficient and also causes an issue if the
object that was serialized is deleted before the job is picked up by a worker
(as it can't then deserialize it).

To avoid this issue we can just pass the `id` of the object, which will take up
much less space in Redis, and load it into memory when the worker picks up the
job. We can guard against the object no longer existing to avoid potential race
conditions/errors.

Add test coverage for analytics upsert jobs.