[2064] Fix declarations api v3 query bug
A LP that a ppt has transferred to or trained with should be able to see historical decs made by another lead provider. This is so they can see how much engagement has taken place.
We have identified a bug where participants with leaving and withdrawn inductions result in providers missing out declarations. Where IR are active these can be found but there's a bug in the service that does not permit this.