Remove uplift caps
NIoT have hit the uplift cap, but according to the contract managers
there never should have been an uplift cap in the first place.

To avoid refactoring a bunch of the finance calculators and their
tests we are just inflating the uplift cap so high that nobody will
ever hit it again.

We won't include uplift caps in RECT.